refactor: adapt assignments + submissions to post-merge conventions
Align with main's policy refactor (c277f9b) and avoid migration number
collisions with main's events schema work:

* Move AssignmentPolicy / SubmissionPolicy into the Policy module
  namespace (Policy::Assignment, Policy::Submission) and rename their
  files + spec files accordingly. Update all 10 service callers. Match
  the self_enrolled? predicate name used by the rest of the policy
  suite.
* Renumber assignments + submissions migrations to 012-015 so they
  follow main's 009-011 events migrations.

Backend: 1166 tests, 0 failures, 0 errors. Frontend prod build clean.
