+++ backend_app/app/application/services/assignments/get_assignment.rb
@@ -32,7 +32,7 @@ module Tyto
 
           @requestor = requestor
           @enrollment = find_enrollment
-          @policy = AssignmentPolicy.new(@requestor, @enrollment)
+          @policy = Policy::Assignment.new(@requestor, @enrollment)
           return Failure(forbidden('You are not authorized to view assignments')) unless @policy.can_view?
 
           load_assignment(assignment_id)
