+++ backend_app/app/application/services/assignments/unpublish_assignment.rb
@@ -32,7 +32,7 @@ module Tyto
 
           @requestor = requestor
           @enrollment = find_enrollment
-          @policy = AssignmentPolicy.new(@requestor, @enrollment)
+          @policy = Policy::Assignment.new(@requestor, @enrollment)
           return Failure(forbidden('You are not authorized to unpublish assignments')) unless @policy.can_unpublish?
 
           load_assignment(assignment_id)
